TASK: Updated composer shell
Check if ./app directory has a composer.json. If so update the app otherwise use default behaviour

+++ bin/composer.sh
@@ -1,2 +1,11 @@
 #!/usr/bin/env bash
-docker-compose run --rm composer $@
+source "./.env"
+
+echo "Checking for app/composer.json"
+if [ -f ./app/composer.json ]
+    then
+        echo 'Updating ./app'
+        docker-compose run --rm composer $@ -d ./app
+    else
+        docker-compose run --rm composer $@
+fi
